# Replace sov