--- a/test/cachedsize/cachedsize.pb.go
+++ b/test/cachedsize/cachedsize.pb.go
@@ -9,6 +9,7 @@ import (
 	_ "github.com/gogo/protobuf/gogoproto"
 	proto "github.com/gogo/protobuf/proto"
 	math "math"
+	math_bits "math/bits"
 )
 
 // Reference imports to suppress errors if they are not otherwise used.
@@ -270,14 +271,7 @@ func (m *Bar) Size() (n int) {
 }
 
 func sovCachedsize(x uint64) (n int) {
-	for {
-		n++
-		x >>= 7
-		if x == 0 {
-			break
-		}
-	}
-	return n
+	return (math_bits.Len64(x|1) + 6) / 7
 }
 func sozCachedsize(x uint64) (n int) {
 	return sovCachedsize(uint64((x << 1) ^ uint64((int64(x) >> 63))))