--- a/test/enumdecl_all/enumdeclall.pb.go
+++ b/test/enumdecl_all/enumdeclall.pb.go
@@ -10,6 +10,7 @@ import (
 	proto "github.com/gogo/protobuf/proto"
 	io "io"
 	math "math"
+	math_bits "math/bits"
 )
 
 // Reference imports to suppress errors if they are not otherwise used.
@@ -357,14 +358,7 @@ func (m *Message) Size() (n int) {
 }
 
 func sovEnumdeclall(x uint64) (n int) {
-	for {
-		n++
-		x >>= 7
-		if x == 0 {
-			break
-		}
-	}
-	return n
+	return (math_bits.Len64(x|1) + 6) / 7
 }
 func sozEnumdeclall(x uint64) (n int) {
 	return sovEnumdeclall(uint64((x << 1) ^ uint64((int64(x) >> 63))))