--- a/test/indeximport-issue72/indeximport.pb.go
+++ b/test/indeximport-issue72/indeximport.pb.go
@@ -11,6 +11,7 @@ import (
 	index "github.com/gogo/protobuf/test/indeximport-issue72/index"
 	io "io"
 	math "math"
+	math_bits "math/bits"
 )
 
 // Reference imports to suppress errors if they are not otherwise used.
@@ -272,14 +273,7 @@ func (m *IndexQueries) Size() (n int) {
 }
 
 func sovIndeximport(x uint64) (n int) {
-	for {
-		n++
-		x >>= 7
-		if x == 0 {
-			break
-		}
-	}
-	return n
+	return (math_bits.Len64(x|1) + 6) / 7
 }
 func sozIndeximport(x uint64) (n int) {
 	return sovIndeximport(uint64((x << 1) ^ uint64((int64(x) >> 63))))