--- a/test/int64support/object.pb.go
+++ b/test/int64support/object.pb.go
@@ -9,6 +9,7 @@ import (
 	proto "github.com/gogo/protobuf/proto"
 	io "io"
 	math "math"
+	math_bits "math/bits"
 	reflect "reflect"
 	strings "strings"
 )
@@ -308,14 +309,7 @@ func (m *Object) Size() (n int) {
 }
 
 func sovObject(x uint64) (n int) {
-	for {
-		n++
-		x >>= 7
-		if x == 0 {
-			break
-		}
-	}
-	return n
+	return (math_bits.Len64(x|1) + 6) / 7
 }
 func sozObject(x uint64) (n int) {
 	return sovObject(uint64((x << 1) ^ uint64((int64(x) >> 63))))