--- a/test/issue260/issue260.pb.go
+++ b/test/issue260/issue260.pb.go
@@ -11,6 +11,7 @@ import (
 	github_com_gogo_protobuf_types "github.com/gogo/protobuf/types"
 	io "io"
 	math "math"
+	math_bits "math/bits"
 	time "time"
 )
 
@@ -644,14 +645,7 @@ func (m *Kept) Size() (n int) {
 }
 
 func sovIssue260(x uint64) (n int) {
-	for {
-		n++
-		x >>= 7
-		if x == 0 {
-			break
-		}
-	}
-	return n
+	return (math_bits.Len64(x|1) + 6) / 7
 }
 func sozIssue260(x uint64) (n int) {
 	return sovIssue260(uint64((x << 1) ^ uint64((int64(x) >> 63))))