--- a/test/issue444/issue444.pb.go
+++ b/test/issue444/issue444.pb.go
@@ -8,6 +8,7 @@ import (
 	_ "github.com/gogo/protobuf/gogoproto"
 	proto "github.com/gogo/protobuf/proto"
 	math "math"
+	math_bits "math/bits"
 )
 
 // Reference imports to suppress errors if they are not otherwise used.
@@ -93,14 +94,7 @@ func (m *SizeMe) Size() (n int) {
 }
 
 func sovIssue444(x uint64) (n int) {
-	for {
-		n++
-		x >>= 7
-		if x == 0 {
-			break
-		}
-	}
-	return n
+	return (math_bits.Len64(x|1) + 6) / 7
 }
 func sozIssue444(x uint64) (n int) {
 	return sovIssue444(uint64((x << 1) ^ uint64((int64(x) >> 63))))