--- a/test/issue449/issue449.pb.go
+++ b/test/issue449/issue449.pb.go
@@ -11,6 +11,7 @@ import (
 	proto "github.com/gogo/protobuf/proto"
 	io "io"
 	math "math"
+	math_bits "math/bits"
 )
 
 // Reference imports to suppress errors if they are not otherwise used.
@@ -346,14 +347,7 @@ func (m *CodeGenMsg) Size() (n int) {
 }
 
 func sovIssue449(x uint64) (n int) {
-	for {
-		n++
-		x >>= 7
-		if x == 0 {
-			break
-		}
-	}
-	return n
+	return (math_bits.Len64(x|1) + 6) / 7
 }
 func sozIssue449(x uint64) (n int) {
 	return sovIssue449(uint64((x << 1) ^ uint64((int64(x) >> 63))))