--- a/test/issue503/issue503.pb.go
+++ b/test/issue503/issue503.pb.go
@@ -10,6 +10,7 @@ import (
 	proto "github.com/gogo/protobuf/proto"
 	io "io"
 	math "math"
+	math_bits "math/bits"
 )
 
 // Reference imports to suppress errors if they are not otherwise used.
@@ -471,14 +472,7 @@ func (m *Foo) Size() (n int) {
 }
 
 func sovIssue503(x uint64) (n int) {
-	for {
-		n++
-		x >>= 7
-		if x == 0 {
-			break
-		}
-	}
-	return n
+	return (math_bits.Len64(x|1) + 6) / 7
 }
 func sozIssue503(x uint64) (n int) {
 	return sovIssue503(uint64((x << 1) ^ uint64((int64(x) >> 63))))