--- a/test/issue530/issue530.pb.go
+++ b/test/issue530/issue530.pb.go
@@ -10,6 +10,7 @@ import (
 	github_com_gogo_protobuf_sortkeys "github.com/gogo/protobuf/sortkeys"
 	io "io"
 	math "math"
+	math_bits "math/bits"
 	reflect "reflect"
 	strings "strings"
 )
@@ -1875,14 +1876,7 @@ func (m *Bar9) Size() (n int) {
 }
 
 func sovIssue530(x uint64) (n int) {
-	for {
-		n++
-		x >>= 7
-		if x == 0 {
-			break
-		}
-	}
-	return n
+	return (math_bits.Len64(x|1) + 6) / 7
 }
 func sozIssue530(x uint64) (n int) {
 	return sovIssue530(uint64((x << 1) ^ uint64((int64(x) >> 63))))