--- a/types/source_context.pb.go
+++ b/types/source_context.pb.go
@@ -9,6 +9,7 @@ import (
 	proto "github.com/gogo/protobuf/proto"
 	io "io"
 	math "math"
+	math_bits "math/bits"
 	reflect "reflect"
 	strings "strings"
 )
@@ -321,14 +322,7 @@ func (m *SourceContext) Size() (n int) {
 }
 
 func sovSourceContext(x uint64) (n int) {
-	for {
-		n++
-		x >>= 7
-		if x == 0 {
-			break
-		}
-	}
-	return n
+	return (math_bits.Len64(x|1) + 6) / 7
 }
 func sozSourceContext(x uint64) (n int) {
 	return sovSourceContext(uint64((x << 1) ^ uint64((int64(x) >> 63))))