… test/mapsproto2/combos/unmarshaler/mapsproto2.pb.go
… test/nopackage/nopackage.pb.go
… test/oneof/combos/both/one.pb.go
… test/oneof/combos/marshaler/one.pb.go
… test/oneof/combos/neither/one.pb.go
… test/oneof/combos/unmarshaler/one.pb.go
… test/oneof3/combos/both/one.pb.go
… test/oneof3/combos/marshaler/one.pb.go
… test/oneof3/combos/neither/one.pb.go
… test/oneof3/combos/unmarshaler/one.pb.go
… test/protosize/protosize.pb.go
… test/required/requiredexample.pb.go
… test/sizeunderscore/sizeunderscore.pb.go
… test/stdtypes/stdtypes.pb.go
… test/theproto3/combos/both/theproto3.pb.go
… test/theproto3/combos/marshaler/theproto3.pb.go
… test/theproto3/combos/neither/theproto3.pb.go
… test/theproto3/combos/unmarshaler/theproto3.pb.go
… test/thetest.pb.go
… test/typedecl/typedecl.pb.go
… test/typedecl_all/typedeclall.pb.go
… test/types/combos/both/types.pb.go
… test/types/combos/marshaler/types.pb.go
… test/types/combos/neither/types.pb.go
… test/types/combos/unmarshaler/types.pb.go
… test/unrecognized/unrecognized.pb.go
… test/unrecognizedgroup/unrecognizedgroup.pb.go
… vanity/test/fast/gogovanity.pb.go
… vanity/test/fast/proto3.pb.go
… vanity/test/fast/vanity.pb.go