Remove the loop in sovXXX by using bit twiddling.

```
benchmark                                                                  old ns/op     new ns/op     delta
BenchmarkNidOptNativeProtoMarshal-12                                       540           469           -13.15%
BenchmarkNinOptNativeProtoMarshal-12                                       535           481           -10.09%
BenchmarkNidRepNativeProtoMarshal-12                                       1492          1310          -12.20%
BenchmarkNinRepNativeProtoMarshal-12                                       1451          1308          -9.86%
BenchmarkNidRepPackedNativeProtoMarshal-12                                 1565          1426          -8.88%
BenchmarkNinRepPackedNativeProtoMarshal-12                                 1609          1456          -9.51%
BenchmarkNidOptStructProtoMarshal-12                                       1434          1279          -10.81%
BenchmarkNinOptStructProtoMarshal-12                                       1345          1158          -13.90%
BenchmarkNidRepStructProtoMarshal-12                                       2987          2541          -14.93%
BenchmarkNinRepStructProtoMarshal-12                                       2718          2336          -14.05%
BenchmarkNidEmbeddedStructProtoMarshal-12                                  1000          904           -9.60%
BenchmarkNinEmbeddedStructProtoMarshal-12                                  962           825           -14.24%
BenchmarkNidNestedStructProtoMarshal-12                                    7535          6258          -16.95%
BenchmarkNinNestedStructProtoMarshal-12                                    6650          5347          -19.59%
BenchmarkNidOptCustomProtoMarshal-12                                       217           225           +3.69%
BenchmarkCustomDashProtoMarshal-12                                         217           203           -6.45%
BenchmarkNinOptCustomProtoMarshal-12                                       219           226           +3.20%
BenchmarkNidRepCustomProtoMarshal-12                                       492           470           -4.47%
BenchmarkNinRepCustomProtoMarshal-12                                       496           464           -6.45%
BenchmarkNinOptNativeUnionProtoMarshal-12                                  133           135           +1.50%
BenchmarkNinOptStructUnionProtoMarshal-12                                  278           271           -2.52%
BenchmarkNinEmbeddedStructUnionProtoMarshal-12                             472           438           -7.20%
BenchmarkNinNestedStructUnionProtoMarshal-12                               361           329           -8.86%
BenchmarkTreeProtoMarshal-12                                               331           325           -1.81%
BenchmarkOrBranchProtoMarshal-12                                           607           595           -1.98%
BenchmarkAndBranchProtoMarshal-12                                          646           550           -14.86%
BenchmarkLeafProtoMarshal-12                                               261           263           +0.77%
BenchmarkDeepTreeProtoMarshal-12                                           455           452           -0.66%
BenchmarkADeepBranchProtoMarshal-12                                        541           549           +1.48%
BenchmarkAndDeepBranchProtoMarshal-12                                      899           863           -4.00%
BenchmarkDeepLeafProtoMarshal-12                                           427           401           -6.09%
BenchmarkNilProtoMarshal-12                                                124           120           -3.23%
BenchmarkNidOptEnumProtoMarshal-12                                         151           151           +0.00%
BenchmarkNinOptEnumProtoMarshal-12                                         189           187           -1.06%
BenchmarkNidRepEnumProtoMarshal-12                                         344           326           -5.23%
BenchmarkNinRepEnumProtoMarshal-12                                         348           309           -11.21%
BenchmarkNinOptEnumDefaultProtoMarshal-12                                  187           189           +1.07%
BenchmarkAnotherNinOptEnumProtoMarshal-12                                  182           191           +4.95%
BenchmarkAnotherNinOptEnumDefaultProtoMarshal-12                           191           190           -0.52%
BenchmarkTimerProtoMarshal-12                                              234           226           -3.42%
BenchmarkMyExtendableProtoMarshal-12                                       428           422           -1.40%
BenchmarkOtherExtenableProtoMarshal-12                                     831           841           +1.20%
BenchmarkNestedDefinitionProtoMarshal-12                                   524           498           -4.96%
BenchmarkNestedDefinition_NestedMessageProtoMarshal-12                     288           294           +2.08%
BenchmarkNestedDefinition_NestedMessage_NestedNestedMsgProtoMarshal-12     212           211           -0.47%
BenchmarkNestedScopeProtoMarshal-12                                        470           437           -7.02%
BenchmarkNinOptNativeDefaultProtoMarshal-12                                513           448           -12.67%
BenchmarkCustomContainerProtoMarshal-12                                    312           290           -7.05%
BenchmarkCustomNameNidOptNativeProtoMarshal-12                             540           442           -18.15%
BenchmarkCustomNameNinOptNativeProtoMarshal-12                             526           473           -10.08%
BenchmarkCustomNameNinRepNativeProtoMarshal-12                             1502          1341          -10.72%
BenchmarkCustomNameNinStructProtoMarshal-12                                1689          1464          -13.32%
BenchmarkCustomNameCustomTypeProtoMarshal-12                               533           551           +3.38%
BenchmarkCustomNameNinEmbeddedStructUnionProtoMarshal-12                   473           435           -8.03%
BenchmarkCustomNameEnumProtoMarshal-12                                     246           239           -2.85%
BenchmarkNoExtensionsMapProtoMarshal-12                                    261           240           -8.05%
BenchmarkUnrecognizedProtoMarshal-12                                       154           152           -1.30%
BenchmarkUnrecognizedWithInnerProtoMarshal-12                              337           317           -5.93%
BenchmarkUnrecognizedWithInner_InnerProtoMarshal-12                        65.6          63.6          -3.05%
BenchmarkUnrecognizedWithEmbedProtoMarshal-12                              274           264           -3.65%
BenchmarkUnrecognizedWithEmbed_EmbeddedProtoMarshal-12                     65.5          61.4          -6.26%
BenchmarkNodeProtoMarshal-12                                               248           269           +8.47%
BenchmarkNonByteCustomTypeProtoMarshal-12                                  251           264           +5.18%
BenchmarkNidOptNonByteCustomTypeProtoMarshal-12                            289           275           -4.84%
BenchmarkNinOptNonByteCustomTypeProtoMarshal-12                            265           263           -0.75%
BenchmarkNidRepNonByteCustomTypeProtoMarshal-12                            631           597           -5.39%
BenchmarkNinRepNonByteCustomTypeProtoMarshal-12                            594           575           -3.20%
BenchmarkProtoTypeProtoMarshal-12                                          205           216           +5.37%

```