@apelisse 2019-04-16
I'm suspecting there is an easy, better way to divide by 7 but I don't have much time to think about it right now.